+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.1.0
+
+* [BUG] add `defaultCondition` to `_getCondition` (Internal).
+* [DEPRECATED] change extension from `isSatisfiedVersion` to `satisfiedWith`.
+* [DEPRECATED] change from `SatisfiedVersion.isSatisfied` to `SatisfiedVersion.string`.
+* Improve test for extension.
+
 ## 0.0.2
 
 * Added support to compare the version within the range for `List`. See README.